CAI2_SentimentAnalysis (1)/CAI2/requirements.txt
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
Flask==3.0.3
Flask-Cors==5.0.0
google-api-core==2.21.0
google-auth==2.35.0
google-cloud-compute==1.19.2
google-cloud-core==2.4.1
google-cloud-functions==1.17.0
google-cloud-language==2.14.0
google-cloud-speech==2.27.0
google-cloud-storage==2.18.2
google-cloud-texttospeech==2.18.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
grpc-google-iam-v1==0.13.1
grpcio==1.67.0
grpcio-status==1.67.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
proto-plus==1.24.0
protobuf==5.28.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
requests==2.32.3
rsa==4.9
urllib3==2.2.3
Werkzeug==3.0.4
gunicorn==20.1.0


